requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
comm==0.1.2
contourpy==1.0.7
customtkinter==5.1.2
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fonttools==4.38.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.1
ipython==8.9.0
jedi==0.18.2
joblib==1.2.0
jupyter-client==8.0.2
jupyter-core==5.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nptyping==2.4.1
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pynrrd==1.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing-extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
wget==3.2
zipp==3.12.0
